+++ oap-server/server-core/src/main/java/org/apache/skywalking/oap/server/core/source/K8SEndpoint.java
@@ -31,7 +31,8 @@ import static org.apache.skywalking.oap.server.core.source.DefaultScopeDefine.K8
 public class K8SEndpoint extends K8SMetrics.ProtocolMetrics {
     private volatile String entityId;
 
-    @ScopeDefaultColumn.DefinedByField(columnName = "service_id", groupByCondInTopN = true)
+    @ScopeDefaultColumn.DefinedByField(columnName = "service_id")
+    @ScopeDefaultColumn.BanyanDB(groupByCondInTopN = true)
     private String serviceId;
     @ScopeDefaultColumn.DefinedByField(columnName = "service_name", requireDynamicActive = true)
     private String serviceName;