+++ oap-server/server-core/src/main/java/org/apache/skywalking/oap/server/core/storage/model/StorageModels.java
@@ -58,7 +58,7 @@ public class StorageModels implements IModelManager, ModelCreator, ModelManipula
         DefaultScopeDefine.nameOf(scopeId);
 
         List<ModelColumn> modelColumns = new ArrayList<>();
-        ShardingKeyChecker checker = new ShardingKeyChecker();
+        SeriesIDChecker checker = new SeriesIDChecker();
         SQLDatabaseModelExtension sqlDBModelExtension = new SQLDatabaseModelExtension();
         BanyanDBModelExtension banyanDBModelExtension = new BanyanDBModelExtension();
         ElasticSearchModelExtension elasticSearchModelExtension = new ElasticSearchModelExtension();
@@ -153,7 +153,7 @@ public class StorageModels implements IModelManager, ModelCreator, ModelManipula
                            final String modelName,
                            final List<ModelColumn> modelColumns,
                            final int scopeId,
-                           ShardingKeyChecker checker,
+                           SeriesIDChecker checker,
                            final SQLDatabaseModelExtension sqlDBModelExtension,
                            final BanyanDBModelExtension banyanDBModelExtension) {
         if (log.isDebugEnabled()) {
@@ -222,13 +222,16 @@ public class StorageModels implements IModelManager, ModelCreator, ModelManipula
                     BanyanDB.TopNAggregation.class);
                 final BanyanDB.MatchQuery analyzer = field.getAnnotation(
                     BanyanDB.MatchQuery.class);
+                final BanyanDB.EnableSort enableSort = field.getAnnotation(
+                    BanyanDB.EnableSort.class);
                 final boolean shouldIndex = (banyanDBNoIndex == null) && !column.storageOnly();
                 BanyanDBExtension banyanDBExtension = new BanyanDBExtension(
                     banyanDBSeriesID == null ? -1 : banyanDBSeriesID.index(),
                     shouldIndex,
                     banyanDBIndexRule == null ? BanyanDB.IndexRule.IndexType.INVERTED : banyanDBIndexRule.indexType(),
                     banyanDBMeasureField != null,
-                    analyzer == null ? null : analyzer.analyzer()
+                    analyzer == null ? null : analyzer.analyzer(),
+                    enableSort != null
                 );
 
                 if (topNAggregation != null) {
@@ -251,7 +254,7 @@ public class StorageModels implements IModelManager, ModelCreator, ModelManipula
                     elasticSearchExtension,
                     banyanDBExtension
                 );
-                if (banyanDBExtension.isShardingKey()) {
+                if (banyanDBExtension.isSeriesID()) {
                     checker.accept(modelName, modelColumn);
                 }
 
@@ -321,14 +324,14 @@ public class StorageModels implements IModelManager, ModelCreator, ModelManipula
         return models;
     }
 
-    private static class ShardingKeyChecker {
+    private static class SeriesIDChecker {
         private final ArrayList<ModelColumn> keys = new ArrayList<>();
 
         /**
-         * @throws IllegalStateException if sharding key indices are conflicting.
+         * @throws IllegalStateException if seriesID indices are conflicting.
          */
         private void accept(String modelName, ModelColumn modelColumn) throws IllegalStateException {
-            final int idx = modelColumn.getBanyanDBExtension().getShardingKeyIdx();
+            final int idx = modelColumn.getBanyanDBExtension().getSeriesIDIdx();
             while (idx + 1 > keys.size()) {
                 keys.add(null);
             }
@@ -337,21 +340,21 @@ public class StorageModels implements IModelManager, ModelCreator, ModelManipula
                 throw new IllegalStateException(
                     modelName + "'s "
                         + "Column [" + exist.getColumnName() + "] and column [" + modelColumn.getColumnName()
-                        + " are conflicting with sharding key index=" + modelColumn.getBanyanDBExtension()
-                                                                                   .getShardingKeyIdx());
+                        + " are conflicting with seriesID index=" + modelColumn.getBanyanDBExtension()
+                                                                                   .getSeriesIDIdx());
             }
             keys.set(idx, modelColumn);
         }
 
         /**
          * @param modelName model name of the entity
-         * @throws IllegalStateException if sharding key indices are not continuous
+         * @throws IllegalStateException if seriesIDs indices are not continuous
          */
         private void check(String modelName) throws IllegalStateException {
             for (int i = 0; i < keys.size(); i++) {
                 final ModelColumn modelColumn = keys.get(i);
                 if (modelColumn == null) {
-                    throw new IllegalStateException("Sharding key index=" + i + " is missing in " + modelName);
+                    throw new IllegalStateException("seriesID index=" + i + " is missing in " + modelName);
                 }
             }
         }